fix imports after moving constants
