+++ BALSAMIC/commands/report/deliver.py
@@ -15,10 +15,10 @@ from BALSAMIC.utils.cli import get_snakefile
 from BALSAMIC.utils.cli import SnakeMake
 from BALSAMIC.utils.cli import convert_deliverables_tags
 from BALSAMIC.utils.rule import get_result_dir
-from BALSAMIC.utils.constants import VCF_DICT
 from BALSAMIC.utils.exc import BalsamicError
 from BALSAMIC.utils.qc_metrics import get_qc_metrics
 from BALSAMIC.utils.qc_report import render_html, report_data_population
+from BALSAMIC.constants.workflow_params import VCF_DICT
 
 LOG = logging.getLogger(__name__)
