+++ BALSAMIC/snakemake_rules/variant_calling/sentieon_tn_varcall.rule
@@ -1,8 +1,6 @@
 # vim: syntax=python tabstop=4 expandtab
 # coding: utf-8
 
-from BALSAMIC.utils.constants import VARCALL_PARAMS
-
 rule sentieon_base_calibration:
     input:
         ref = config["reference"]["reference_genome"],
