+++ BALSAMIC/utils/rule.py
@@ -6,7 +6,7 @@ from pathlib import Path
 import snakemake
 from BALSAMIC.utils.cli import get_file_extension
 from BALSAMIC.utils.cli import find_file_index
-from BALSAMIC.utils.constants import (
+from BALSAMIC.constants.common import (
     MUTATION_TYPE,
     MUTATION_CLASS,
     SEQUENCING_TYPE,
